Fix some lazy loading and lack of highlighting (telescope-fzf-native)
After disabling and enabling plugins "binary search" style for quite a bit I landed on this:
I was ending up without syntax highlighting because of telescope-fzf-native. I was lazy loading it on BufReadPre which apparently is not a great idea and really just not the right way of lazy loading something like that.

Fixed by actually just making the fzf-native extension load `after`
Telescope explicitly while still having telescope itself be lazy loaded.