fix: no highlighting on first file open
The issue was due to lazy loaded vsnip on `BufReadPre`. It is assumed it
was causing some sort of timeout on reading the buffer and causing the
rest of the normal operations when reading a buffer to not occur, like
syntax highlighting. When the file was reloaded, it would then work
since the snippets had already been loaded.